api/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
MarkupSafe==2.1.3
urllib3==1.26.16
PyJWT==2.7.0
